+++ libsmx/smx-builder.cpp
@@ -34,8 +34,8 @@ SmxBuilder::write(ISmxBuffer* buf)
   header.version = SmxConsts::SP1_VERSION_1_1;
   header.compression = SmxConsts::FILE_COMPRESSION_NONE;
 
-  header.disksize = sizeof(header) +
-                    sizeof(sp_file_section_t) * sections_.size();
+  header.disksize = (uint32_t)(sizeof(header) +
+                    sizeof(sp_file_section_t) * sections_.size());
 
   // Note that |dataoffs| here is just to mimic what it would be in earlier
   // versions of Pawn. Its value does not actually matter per the SMX spec,
@@ -46,17 +46,17 @@ SmxBuilder::write(ISmxBuffer* buf)
   size_t current_string_offset = 0;
   for (size_t i = 0; i < sections_.size(); i++) {
     RefPtr<SmxSection> section = sections_[i];
-    header.disksize += section->length();
+    header.disksize += (uint32_t)section->length();
     current_string_offset += section->name().size() + 1;
   }
-  header.disksize += current_string_offset;
-  header.dataoffs += current_string_offset;
+  header.disksize += (uint32_t)current_string_offset;
+  header.dataoffs += (uint32_t)current_string_offset;
 
   header.imagesize = header.disksize;
-  header.sections = sections_.size();
+  header.sections = (uint8_t)sections_.size();
 
   // We put the string table after the sections table.
-  header.stringtab = sizeof(header) + sizeof(sp_file_section_t) * sections_.size();
+  header.stringtab = (uint32_t)(sizeof(header) + sizeof(sp_file_section_t) * sections_.size());
 
   if (!buf->write(&header, sizeof(header)))
     return false;
@@ -68,9 +68,9 @@ SmxBuilder::write(ISmxBuffer* buf)
   current_string_offset = 0;
   for (size_t i = 0; i < sections_.size(); i++) {
     sp_file_section_t s;
-    s.nameoffs = current_string_offset;
-    s.dataoffs = current_data_offset;
-    s.size = sections_[i]->length();
+    s.nameoffs = (uint32_t)current_string_offset;
+    s.dataoffs = (uint32_t)current_data_offset;
+    s.size = (uint32_t)sections_[i]->length();
     if (!buf->write(&s, sizeof(s)))
       return false;
